Better customisation of matplotlib

+++ generate_tables.py
@@ -3,8 +3,8 @@ import numpy as np
 import math
 
 
-# Cleans all the entries with binary strings on file 'file' and 
-# column 'columnWithBinaries'
+# Cleans all the entries with binary strings on file "file" and 
+# column "columnWithBinaries"
 def cleanBinaries(file, columnWithBinaries):
     L = []
     for _, row in file.iterrows():
@@ -14,8 +14,8 @@ def cleanBinaries(file, columnWithBinaries):
     newTable = pd.DataFrame(L)
     return newTable
 
-# Creates columns from column 'column' from file 'file', where the columns have 
-# the names 'names'
+# Creates columns from column "column" from file "file", where the columns have 
+# the names "names"
 def createColumnsWithValues(file, column, names):
     L = [names]
     for _, row in file.iterrows():
@@ -27,9 +27,9 @@ def createColumnsWithValues(file, column, names):
     newTable = pd.concat([file.reset_index(drop=True), LToDataframe.reset_index(drop=True)], axis=1, sort=False)
     return newTable
 
-# Creates columns from column 'column' from file 'file', where the columns have 
-# the names 'names', where the value in the row of the new column will be 0 if the 
-# corresponding value on column 'column' is equal to 'valueCondition, otherwise 1
+# Creates columns from column "column" from file "file", where the columns have 
+# the names "names", where the value in the row of the new column will be 0 if the 
+# corresponding value on column "column" is equal to "valueCondition", otherwise 1
 def createColumnsWithValuesCondition(file, column, valueCondition, names):
     L = [names]
     for _, row in file.iterrows():
@@ -41,7 +41,7 @@ def createColumnsWithValuesCondition(file, column, valueCondition, names):
     newTable = pd.concat([file.reset_index(drop=True), LToDataframe.reset_index(drop=True)], axis=1, sort=False)
     return newTable
 
-# Creates a table from file 'file', where column 'column' has to be a digit
+# Creates a table from file "file", where column "column" has to be a digit
 def createTableForHistograms(file, column):
     L = []
     for _, row in file.iterrows():
@@ -54,41 +54,41 @@ def createTableForHistograms(file, column):
 # Changes the types of data in the columns
 def changeTypeOfData(file):
     newTable = file
-    newTable['coord_lat'] = newTable['coord_lat'].str.replace(',', '.').astype(float)
-    newTable['coord_lon'] = newTable['coord_lon'].str.replace(',', '.').astype(float)
-    newTable['carriage_reader'] = newTable['carriage_reader'].astype(str)
-    newTable['carriage_source'] = newTable['carriage_source'].astype(str)
-    newTable['event_type'] = newTable['event_type'].astype(str)
-    newTable['error_type'] = newTable['error_type'].str.replace('\\N', '-1', regex=False).astype(int)
-    tableWithValues['Corrente eléctrica moderável (mA)'] = tableWithValues['Corrente eléctrica moderável (mA)'].astype(float)
-    tableWithValues['Velocidade de referência (km/h)'] = tableWithValues['Velocidade de referência (km/h)'].astype(float)
-    tableWithValues['Velocidade real eixo 3 (km/h)'] = tableWithValues['Velocidade real eixo 3 (km/h)'].astype(float)
-    tableWithValues['Velocidade real eixo 4 (km/h)'] = tableWithValues['Velocidade real eixo 4 (km/h)'].astype(float)
-    tableWithValues['Velocidade real eixo 1 (km/h)'] = tableWithValues['Velocidade real eixo 1 (km/h)'].astype(float)
-    tableWithValues['Velocidade real eixo 2 (km/h)'] = tableWithValues['Velocidade real eixo 2 (km/h)'].astype(float)
-    tableWithValues['Comando esforço (kN)'] = tableWithValues['Comando esforço (kN)'].astype(float)
-    tableWithValues['Esforço (real) (kN)'] = tableWithValues['Esforço (real) (kN)'].astype(float)
-    tableWithValues['Corrente Conversor 4Q (Aeff)'] = tableWithValues['Corrente Conversor 4Q (Aeff)'].astype(float)
-    tableWithValues['Tensão Catenária (Veff)'] = tableWithValues['Tensão Catenária (Veff)'].astype(float)
+    newTable["coord_lat"] = newTable["coord_lat"].str.replace(',', '.').astype(float)
+    newTable["coord_lon"] = newTable["coord_lon"].str.replace(',', '.').astype(float)
+    newTable["carriage_reader"] = newTable["carriage_reader"].astype(str)
+    newTable["carriage_source"] = newTable["carriage_source"].astype(str)
+    newTable["event_type"] = newTable["event_type"].astype(str)
+    newTable["error_type"] = newTable["error_type"].str.replace('\\N', '-1', regex=False).astype(int)
+    tableWithValues["Corrente eléctrica moderável (mA)"] = tableWithValues["Corrente eléctrica moderável (mA)"].astype(float)
+    tableWithValues["Velocidade de referência (km/h)"] = tableWithValues["Velocidade de referência (km/h)"].astype(float)
+    tableWithValues["Velocidade real eixo 3 (km/h)"] = tableWithValues["Velocidade real eixo 3 (km/h)"].astype(float)
+    tableWithValues["Velocidade real eixo 4 (km/h)"] = tableWithValues["Velocidade real eixo 4 (km/h)"].astype(float)
+    tableWithValues["Velocidade real eixo 1 (km/h)"] = tableWithValues["Velocidade real eixo 1 (km/h)"].astype(float)
+    tableWithValues["Velocidade real eixo 2 (km/h)"] = tableWithValues["Velocidade real eixo 2 (km/h)"].astype(float)
+    tableWithValues["Comando esforço (kN)"] = tableWithValues["Comando esforço (kN)"].astype(float)
+    tableWithValues["Esforço (real) (kN)"] = tableWithValues["Esforço (real) (kN)"].astype(float)
+    tableWithValues["Corrente Conversor 4Q (Aeff)"] = tableWithValues["Corrente Conversor 4Q (Aeff)"].astype(float)
+    tableWithValues["Tensão Catenária (Veff)"] = tableWithValues["Tensão Catenária (Veff)"].astype(float)
     return newTable
 
-path = 'C:/Users/Pedro Trabuco/Documents/Universidade/5º Ano/Tese/code/tables/'
+path = "C:/Users/Pedro Trabuco/Documents/Universidade/5º Ano/Tese/code/tables/"
 
 # Reads the data from the .csv file and drops unwanted/useless columns
-vals = pd.read_csv('exported_csv_commas_full_data.csv', encoding='utf_8', sep=';')
+vals = pd.read_csv("exported_csv_commas_full_data.csv", encoding="utf_8", sep=';')
 print(vals.head())
-vals.drop(['error_sub_type'], axis=1)
-vals.drop(['date_train'], axis=1)
+vals.drop(["error_sub_type"], axis=1)
+vals.drop(["date_train"], axis=1)
 print("Number of entries on original table: {}".format(len(vals.index)))
 
-# Creates copy of the table 'vals' without the entries with binary numbers in 
-# the 'value' column
-tableWithoutBinaries = cleanBinaries(vals, 'value')
+# Creates copy of the table "vals" without the entries with binary numbers in 
+# the "value" column
+tableWithoutBinaries = cleanBinaries(vals, "value")
 print(tableWithoutBinaries.head())
 print("Number of entries on cleaned table: {}".format(len(tableWithoutBinaries.index)))
-tableWithoutBinaries.to_csv(path + "withoutBinaries.csv", encoding='utf-8', index=False)
+tableWithoutBinaries.to_csv(path + "withoutBinaries.csv", encoding="utf-8", index=False)
 
-# Creates columns with values from column 'value', with the given names
+# Creates columns with values from column "value", with the given names
 names = ["Corrente eléctrica moderável (mA)", "Velocidade de referência (km/h)", "Velocidade real eixo 3 (km/h)", \
     "Velocidade real eixo 4 (km/h)", "Velocidade real eixo 1 (km/h)", "Velocidade real eixo 2 (km/h)", \
     "Comando esforço (kN)", "Esforço (real) (kN)", "Corrente Conversor 4Q (Aeff)", "Tensão Catenária (Veff)"]
@@ -96,7 +96,7 @@ tableWithValues = createColumnsWithValues(tableWithoutBinaries, 0, names)
 print(tableWithValues.head())
 print(tableWithValues.tail())
 print("Number of entries on table with values: {}".format(len(tableWithValues.index)))
-tableWithValues.to_csv(path + "tableWithValues.csv", encoding='utf-8', index=False)
+tableWithValues.to_csv(path + "tableWithValues.csv", encoding="utf-8", index=False)
 
 # Changes the types of data and drops irrelevant columns
 print(tableWithValues.dtypes)
@@ -104,16 +104,16 @@ tableWithValuesAndTypes = changeTypeOfData(tableWithValues)
 print(tableWithValuesAndTypes.head())
 print(tableWithValuesAndTypes.tail())
 print(tableWithValuesAndTypes.dtypes)
-tableWithValuesAndTypes.to_csv(path + "tableWithValuesAndTypes.csv", encoding='utf-8', index=False)
+tableWithValuesAndTypes.to_csv(path + "tableWithValuesAndTypes.csv", encoding="utf-8", index=False)
 
 # Creates table without rows without values in the most right columns
 tableForHistograms = createTableForHistograms(tableWithValuesAndTypes, 13)
 print(tableForHistograms.head())
 print(tableForHistograms.tail())
-tableForHistograms.to_csv(path + "tableForHistograms.csv", encoding='utf-8', index=False)
+tableForHistograms.to_csv(path + "tableForHistograms.csv", encoding="utf-8", index=False)
 
 # Creates a column where every row has 0 if there isn't an error, 1 otherwise
-tableWithErrorColumn = createColumnsWithValuesCondition(tableWithValuesAndTypes, 11, -1, ['Error'])
+tableWithErrorColumn = createColumnsWithValuesCondition(tableWithValuesAndTypes, 11, -1, ["Error"])
 print(tableWithErrorColumn.head())
 print(tableWithErrorColumn.tail())
-tableWithErrorColumn.to_csv(path + "tableWithErrorColumn.csv", encoding='utf-8', index=False)
+tableWithErrorColumn.to_csv(path + "tableWithErrorColumn.csv", encoding="utf-8", index=False)
